// apps/dokument-app/src/api/BidragDokumentApi.ts
export interface AktorDto {
    /** Identifaktor til aktøren */
    ident: string;
    /** Identtypene til en aktør */
    type?: IdentType;
}

/**
 *
 * Avsender journalposten ble sendt fra hvis utgående.
 * Mottaker journalposten skal sendes til hvis inngående.
 */
export interface AvsenderMottakerDto {
    /** Avsenders/Mottakers navn (med eventuelt fornavn bak komma). Skal ikke oppgis hvis ident er en FNR */
    navn?: string;
    /** Person ident eller organisasjonsnummer */
    ident?: string;
    /** Identtype */
    type: AvsenderMottakerDtoIdType;
    /** Adresse til mottaker hvis dokumentet skal sendes/er sendt gjennom sentral print */
    adresse?: MottakerAdresseTo;
}

/** Identtype */
export enum AvsenderMottakerDtoIdType {
    FNR = "FNR",
    SAMHANDLER = "SAMHANDLER",
    ORGNR = "ORGNR",
    UTENLANDSK_ORGNR = "UTENLANDSK_ORGNR",
    UKJENT = "UKJENT",
}

/** Identtypene til en aktør */
export enum IdentType {
    AKTOERID = "AKTOERID",
    FNR = "FNR",
    ORGNR = "ORGNR",
}
// ...
export interface MottakerAdresseTo {
    adresselinje1: string;
    adresselinje2?: string;
    adresselinje3?: string;
    bruksenhetsnummer?: string;
    /** Lankode må være i ISO 3166-1 alpha-2 format */
    landkode?: string;
    /** Lankode må være i ISO 3166-1 alpha-3 format */
    landkode3?: string;
    postnummer?: string;
    poststed?: string;
}
// ...
export interface DistribuerTilAdresse {
    adresselinje1?: string;
    adresselinje2?: string;
    adresselinje3?: string;
    /** ISO 3166-1 alpha-2 to-bokstavers landkode */
    land?: string;
    postnummer?: string;
    poststed?: string;
}

/** Arkivsystem hvor dokumentet er produsert og lagret */
export enum DokumentArkivSystemDto {
    JOARK = "JOARK",
    MIDLERTIDLIG_BREVLAGER = "MIDLERTIDLIG_BREVLAGER",
    UKJENT = "UKJENT",
    BIDRAG = "BIDRAG",
    FORSENDELSE = "FORSENDELSE",
}

/** Metadata for et dokument */
export interface DokumentDto {
    /** Referansen til dokumentet i arkivsystemet */
    dokumentreferanse?: string;
    /** Journalpost hvor dokumentet er arkivert. Dette brukes hvis dokumentet er arkivert i annen arkivsystem enn det som er sendt med i forespørsel. */
    journalpostId?: string;
    /**
     * Inngående (I), utgående (U) dokument, (X) internt notat
     * @deprecated
     */
    dokumentType?: string;
    /** Kort oppsummering av dokumentets innhold */
    tittel?: string;
    /** Selve PDF dokumentet formatert som Base64 */
    dokument?: string;
    /**
     * Typen dokument. Brevkoden sier noe om dokumentets innhold og oppbygning. Erstattes av dokumentmalId
     * @deprecated
     */
    brevkode?: string;
    /** Typen dokument. Dokumentmal sier noe om dokumentets innhold og oppbygning. */
    dokumentmalId?: string;
    /** Dokumentets status. Benyttes hvis journalposten er av typen forsendelse */
    status?: DokumentStatusDto;
    /** Arkivsystem hvor dokumentet er produsert og lagret */
    arkivSystem?: DokumentArkivSystemDto;
    /** Metadata om dokumentet */
    metadata: Record<string, string>;
}

/** Dokumentets status. Benyttes hvis journalposten er av typen forsendelse */
export enum DokumentStatusDto {
    IKKE_BESTILT = "IKKE_BESTILT",
    BESTILLING_FEILET = "BESTILLING_FEILET",
    UNDER_PRODUKSJON = "UNDER_PRODUKSJON",
    UNDER_REDIGERING = "UNDER_REDIGERING",
    FERDIGSTILT = "FERDIGSTILT",
    AVBRUTT = "AVBRUTT",
}
// ...
export interface EttersendingsoppgaveVedleggDto {
    tittel?: string;
    vedleggsnr: string;
    status: EttersendingsoppgaveVedleggDtoStatusEnum;
}

export interface EttersendingsppgaveDto {
    tittel: string;
    skjemaId: string;
    innsendingsId?: string;
    språk: string;
    status: EttersendingsppgaveDtoStatusEnum;
    /** @format date */
    opprettetDato?: string;
    /** @format date */
    fristDato?: string;
    /** @format date */
    slettesDato?: string;
    vedleggsliste: EttersendingsoppgaveVedleggDto[];
}

/** Metadata til en journalpost */
export interface JournalpostDto {
    /**
     * Avsenders navn (med eventuelt fornavn bak komma)
     * @deprecated
     */
    avsenderNavn?: string;
    /**
     *
     * Avsender journalposten ble sendt fra hvis utgående.
     * Mottaker journalposten skal sendes til hvis inngående.
     */
    avsenderMottaker?: AvsenderMottakerDto;
    /** Dokumentene som følger journalposten */
    dokumenter: DokumentDto[];
    /**
     * Dato for dokument i journalpost
     * @format date
     */
    dokumentDato?: string;
    /**
     * Tidspunkt for dokument i journalpost
     * @format date-time
     */
    dokumentTidspunkt?: string;
    /**
     * Dato dokumentene på journalposten ble sendt til bruker.
     * @format date
     */
    ekspedertDato?: string;
    /** Fagområdet for journalposten. BID for bidrag og FAR for farskap */
    fagomrade?: string;
    /** Ident for hvem/hva dokumente(t/ne) gjelder */
    gjelderIdent?: string;
    /** Metadata om en aktør */
    gjelderAktor?: AktorDto;
    /** Kort oppsummert av journalført innhold */
    innhold?: string;
    /** Enhetsnummer hvor journalføring ble gjort */
    journalforendeEnhet?: string;
    /** Saksbehandler som var journalfører */
    journalfortAv?: string;
    /**
     * Dato dokument ble journalført
     * @format date
     */
    journalfortDato?: string;
    /** Identifikator av journalpost i midlertidig brevlager eller fra joark på formatet [BID|JOARK]-<journalpostId> */
    journalpostId?: string;
    /** Journalposten ble mottatt/sendt ut i kanal */
    kilde?: Kanal;
    /** Journalposten ble mottatt/sendt ut i kanal */
    kanal?: Kanal;
    /**
     * Dato for når dokument er mottat, dvs. dato for journalføring eller skanning
     * @format date
     */
    mottattDato?: string;
    /** Inngående (I), utgående (U) journalpost; (X) internt notat */
    dokumentType?: string;
    /**
     * Journalpostens status, (A, D, J, M, O, R, S, T, U, KP, EJ, E)
     * @deprecated
     */
    journalstatus?: string;
    /** Journalpostens status */
    status?: JournalpostStatus;
    /** Om journalposten er feilført på bidragssak */
    feilfort?: boolean;
    /** Metadata for kode vs dekode i et kodeobjekt */
    brevkode?: KodeDto;
    /** Metadata for retur detaljer */
    returDetaljer?: ReturDetaljer;
    /** Joark journalpostid for bidrag journalpost som er arkivert i Joark */
    joarkJournalpostId?: string;
    /** Adresse for hvor brev sendes ved sentral print */
    distribuertTilAdresse?: DistribuerTilAdresse;
    /** Informasjon om returdetaljer til journalpost */
    sakstilknytninger: string[];
    /** Språket til dokumentet i Journalposten */
    språk?: string;
    /** Saksbehandler som opprettet journalposten */
    opprettetAvIdent?: string;
    /** Referanse til originale kilden til journalposten. Kan være referanse til forsendelse eller bidrag journalpost med prefiks. Feks BID_12323 eller BIF_123213 */
    eksternReferanseId?: string;
    ettersendingsppgave?: EttersendingsppgaveDto;
}

/** Journalpostens status */
export enum JournalpostStatus {
    AVVIK_ENDRE_FAGOMRADE = "AVVIK_ENDRE_FAGOMRADE",
    AVVIK_BESTILL_RESKANNING = "AVVIK_BESTILL_RESKANNING",
    AVVIK_BESTILL_SPLITTING = "AVVIK_BESTILL_SPLITTING",
    MOTTATT = "MOTTATT",
    JOURNALFORT = "JOURNALFØRT",
    EKSPEDERT = "EKSPEDERT",
    EKSPEDERT_JOARK = "EKSPEDERT_JOARK",
    MOTTAKSREGISTRERT = "MOTTAKSREGISTRERT",
    UKJENT = "UKJENT",
    DISTRIBUERT = "DISTRIBUERT",
    AVBRUTT = "AVBRUTT",
    KLAR_FOR_DISTRIBUSJON = "KLAR_FOR_DISTRIBUSJON",
    RETUR = "RETUR",
    FERDIGSTILT = "FERDIGSTILT",
    FEILREGISTRERT = "FEILREGISTRERT",
    RESERVERT = "RESERVERT",
    UTGAR = "UTGÅR",
    SLETTET = "SLETTET",
    UNDER_OPPRETTELSE = "UNDER_OPPRETTELSE",
    TIL_LAGRING = "TIL_LAGRING",
    OPPRETTET = "OPPRETTET",
    UNDER_PRODUKSJON = "UNDER_PRODUKSJON",
}

/** Journalposten ble mottatt/sendt ut i kanal */
export enum Kanal {
    NAV_NO = "NAV_NO",
    NAV_NO_BID = "NAV_NO_BID",
    SKAN_BID = "SKAN_BID",
    SKAN_NETS = "SKAN_NETS",
    SKAN_IM = "SKAN_IM",
    LOKAL_UTSKRIFT = "LOKAL_UTSKRIFT",
    SENTRAL_UTSKRIFT = "SENTRAL_UTSKRIFT",
    SDP = "SDP",
    INGEN_DISTRIBUSJON = "INGEN_DISTRIBUSJON",
    INNSENDT_NAV_ANSATT = "INNSENDT_NAV_ANSATT",
    NAV_NO_UINNLOGGET = "NAV_NO_UINNLOGGET",
    NAV_NO_CHAT = "NAV_NO_CHAT",
}

/** Metadata for kode vs dekode i et kodeobjekt */
export interface KodeDto {
    /** Koden */
    kode?: string;
    /** Dekode (kodebeskrivelse) */
    dekode?: string;
    /** Om kodeobjektet inneholder en gyldig verdi */
    erGyldig: boolean;
}

/** Metadata for retur detaljer */
export interface ReturDetaljer {
    /**
     * Dato for siste retur
     * @format date
     */
    dato?: string;
    /**
     * Totalt antall returer
     * @format int32
     */
    antall?: number;
    /** Liste med logg av alle registrerte returer */
    logg: ReturDetaljerLog[];
}

/** Metadata for retur detaljer log */
export interface ReturDetaljerLog {
    /**
     * Dato for retur
     * @format date
     */
    dato?: string;
    /** Beskrivelse av retur (eks. addresse forsøkt sendt) */
    beskrivelse?: string;
    /** Returdetalje er låst for endring. Dette blir satt etter en ny distribusjon er bestilt */
    locked?: boolean;
}
// ...
export enum EttersendingsppgaveDtoStatusEnum {
    OPPRETTET = "OPPRETTET",
    UTFYLT = "UTFYLT",
    INNSENDT = "INNSENDT",
    SLETTET_AV_BRUKER = "SLETTET_AV_BRUKER",
    AUTOMATISK_SLETTET = "AUTOMATISK_SLETTET",
    UKJENT = "UKJENT",
    IKKE_OPPRETTET = "IKKE_OPPRETTET",
}